+++ src/lib/pipeline/BalancedPortAllocator.cc
@@ -11,12 +11,12 @@ BalancedPortAllocator::BalancedPortAllocator(
   // Construct the  support matrix
   for (size_t portIndex = 0; portIndex < portArrangement.size(); portIndex++) {
     const auto& port = portArrangement[portIndex];
-    uint8_t id = 0;
+    uint16_t id = 0;
     // Add this port to the matrix entry for each group it supports
     for (const auto& set : port) {
-      std::vector<uint8_t> acceptedSelection;
-      uint8_t compulsoryGroups = 0;
-      std::vector<uint8_t> optionalGroups;
+      std::vector<uint16_t> acceptedSelection;
+      uint16_t compulsoryGroups = 0;
+      std::vector<uint16_t> optionalGroups;
       for (const auto& group : set) {
         assert(group.second < 2 && "port type not supported");
         if (group.second == PortType::COMPULSORY) {
@@ -41,9 +41,9 @@ BalancedPortAllocator::BalancedPortAllocator(
       supportMatrix[compulsoryGroups].push_back(portIndex);
       int n = 0;
       while (n < optionalGroups.size()) {
-          std::vector<uint8_t> temp = acceptedSelection;
+          std::vector<uint16_t> temp = acceptedSelection;
           for (const auto& entry : temp) {
-            uint8_t groupSet = entry | optionalGroups[n];
+            uint16_t groupSet = entry | optionalGroups[n];
             acceptedSelection.push_back(groupSet);
             if (groupSet >= supportMatrix.size()) {
               // New highest group ID; expand matrix
